Fix error when the temp version string file doesn't yet exist.

+++ SetVersionString.pl
@@ -10,7 +10,9 @@ use strict;
 
 die "$0: Must be run from Xcode" unless $ENV{"BUILT_PRODUCTS_DIR"};
 
-my $deleted = `rm "$ENV{TEMP_VERSION_STRING_PATH}"`;
+if (-e $ENV{TEMP_VERSION_STRING_PATH}) {
+	my $deleted = `rm "$ENV{TEMP_VERSION_STRING_PATH}"`;
+}
 
 my $COMMITS_SINCE_LAST_TAG=`git log \`git rev-list --tags --abbrev=0 --max-count=1\`.. --pretty=format:'%h' | wc -l | sed 's/[ \t]//g'`;
 my $NEW_VERSION = $ENV{"VERSION_STRING"};
